+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **New Feature**
   * Added `back` and `forward` methods for easy history manipulation
   * Added `beforeroute` event for manipulating the middleware arguments, so `route` event handlers can can assume the arguments are complete
+  * Added `beforeExit` for registering handlers that can prevent navigation
 * **Breaking Change**
   * Moved `routeLinks` option from `start` method to `Router` construction
   * Changed `routeLinks` to default to `true`, you must pass `routeLinks: false` to disable
@@ -13,7 +14,7 @@
   * Changes `navigate` to return a promise that resolves when finished routing the new path
   * Changed `replace` to return undefined, since no routing is performed
 * **Internal**
-  * Moved all event listening / client-only code to history.js
+  * Moved all event listening / client-only code to it's own file
   * Drop testing IE10
   * Add testing Safari 8 & 9